+++ Simulation/Aivika/Internal/Event.hs
@@ -36,6 +36,7 @@ module Simulation.Aivika.Internal.Event
         enqueueEventWithTimes,
         enqueueEventWithPoints,
         enqueueEventWithIntegTimes,
+        enqueueEventWithStopTime,
         yieldEvent,
         eventQueueCount,
         -- * Cancelling Event
@@ -320,6 +321,13 @@ enqueueEventWithIntegTimes e =
   Event $ \p ->
   let points = integPointsStartingFrom p
   in invokeEvent p $ enqueueEventWithPoints points e
+
+-- | Actuate the event handler in the final time point.
+enqueueEventWithStopTime :: Event () -> Event ()
+enqueueEventWithStopTime e =
+  Event $ \p ->
+  let p0 = integStopPoint $ pointRun p
+  in invokeEvent p $ enqueueEventWithPoints [p0] e
 
 -- | It allows cancelling the event.
 data EventCancellation =