+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Version 4.3.6
 
 * The Transform computation seems to be not ArrowLoop.
 
+* Added the enqueueEventWithStopTime function.
+
 Version 4.3.5
 -----